Multiple game balance Iterations made and implemented along with premium currency and background pricing.
changed wave spawn interval value to be a float and changed UI elements linked with Unit pricing and cash shop pricing.

+++ Assets/Scripts/Gameplay/WaveSpawner.cs
@@ -7,7 +7,7 @@ using Photon.Pun;
 public struct WaveSpawnInfo
 {
     public int m_count;             // Amount of common monsters to spawn
-    public int m_spawnInterval;     // Interval for spawning a new monster
+    public float m_spawnInterval;     // Interval for spawning a new monster
 }
 
 public class WaveSpawner : MonoBehaviourPun
